+++ src/core/layers/SpanLayer.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 import * as PureRender from 'pure-render-decorator';
 import * as d3Scale from 'd3-scale';
 import * as _ from 'lodash';
+import { deprecate } from 'react-is-deprecated';
 
 import NonReactRender from '../decorators/NonReactRender';
 import PixelRatioContext, { Context } from '../decorators/PixelRatioContext';
@@ -11,14 +12,20 @@ import { getIndexBoundsForSpanData } from '../renderUtils';
 import propTypes from '../propTypes';
 import { Interval, Color } from '../interfaces';
 
+export interface ColoredSpanDatum {
+  minXValue: number;
+  maxXValue: number;
+  color?: Color;
+  fillColor?: Color;
+  borderColor?: Color;
+}
+
 export interface Props {
-  data: {
-    minXValue: number;
-    maxXValue: number;
-    color?: Color;
-  }[];
+  data: ColoredSpanDatum[];
   xDomain: Interval;
   color?: Color;
+  fillColor?: Color;
+  borderColor?: Color;
 }
 
 @PureRender
@@ -31,14 +38,18 @@ export default class SpanLayer extends React.Component<Props, void> {
     data: React.PropTypes.arrayOf(React.PropTypes.shape({
       minXValue: React.PropTypes.number.isRequired,
       maxXValue: React.PropTypes.number.isRequired,
-      color: React.PropTypes.string
+      color: deprecate(React.PropTypes.string, 'SpanLayer\'s \'data[].color\' prop is deprecated in favor of \'data[].fillColor\' and/or \'data[].borderColor\''),
+      fillColor: React.PropTypes.string,
+      borderColor: React.PropTypes.string
     })).isRequired,
     xDomain: propTypes.interval.isRequired,
-    color: React.PropTypes.string
+    color: deprecate(React.PropTypes.string, 'SpanLayer\'s \'color\' prop is deprecated in favor of \'fillColor\' and/or \'borderColor\''),
+    fillColor: React.PropTypes.string,
+    borderColor: React.PropTypes.string
   };
 
   static defaultProps = {
-    color: 'rgba(0, 0, 0, 0.1)'
+    fillColor: 'rgba(0, 0, 0, 0.1)'
   };
 
   render() {
@@ -51,23 +62,42 @@ export default class SpanLayer extends React.Component<Props, void> {
 
   nonReactRender = () => {
     const { width, height, context } = (this.refs['canvasLayer'] as PollingResizingCanvasLayer).resetCanvas();
+    _renderCanvas(this.props, width, height, context);
+  }
+}
 
-    const { firstIndex, lastIndex } = getIndexBoundsForSpanData(this.props.data, this.props.xDomain, 'minXValue', 'maxXValue');
-    if (firstIndex === lastIndex) {
-      return;
+// Export for testing.
+export function _renderCanvas(props: Props, width: number, height: number, context: CanvasRenderingContext2D) {
+  const { firstIndex, lastIndex } = getIndexBoundsForSpanData(props.data, props.xDomain, 'minXValue', 'maxXValue');
+  if (firstIndex === lastIndex) {
+    return;
+  }
+
+  const xScale = d3Scale.scaleLinear()
+    .domain([ props.xDomain.min, props.xDomain.max ])
+    .rangeRound([ 0, width ]);
+
+  const defaultFill = props.color || props.fillColor;
+  const defaultStroke = props.borderColor;
+
+  context.lineWidth = 1;
+
+  for (let i = firstIndex; i < lastIndex; ++i) {
+    const left = xScale(props.data[i].minXValue);
+    const right = xScale(props.data[i].maxXValue);
+    context.beginPath();
+    context.rect(left, -1, right - left, height + 2);
+
+    const strokeStyle = props.data[i].borderColor || defaultStroke;
+    if (strokeStyle) {
+      context.strokeStyle = strokeStyle;
+      context.stroke();
     }
 
-    const xScale = d3Scale.scaleLinear()
-      .domain([ this.props.xDomain.min, this.props.xDomain.max ])
-      .rangeRound([ 0, width ]);
-
-    for (let i = firstIndex; i < lastIndex; ++i) {
-      const left = xScale(this.props.data[i].minXValue);
-      const right = xScale(this.props.data[i].maxXValue);
-      context.beginPath();
-      context.rect(left, 0, right - left, height);
-      context.fillStyle = this.props.data[i].color || this.props.color;
+    const fillStyle = props.data[i].fillColor || defaultFill;
+    if (fillStyle) {
+      context.fillStyle = fillStyle;
       context.fill();
     }
-  };
+  }
 }