+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Non-standard `npm run dev:hot-reload` replaced with standard `npm start`. [#33](https://github.com/palantir/react-layered-chart/pull/33).
 - `mergeIntervals` accepts a default argument to return in the case when no intervals are given. [#20](https://github.com/palantir/react-layered-chart/issues/20).
 - `HoverLineLayer`'s `stroke` prop deprecated in favor of `color`.
+- `SpanLayer`'s `color` prop deprecated in favor of `fillColor` and `borderColor`.
 
 ## 1.1.0 (2016-06-09)